Add support for quering milliseconds from TSDB gui.
Gnuplot does not support milliseconds so the best alternative
was to keep plotting data points in the graph as seconds.

Timestamps in the GUI are treated as seconds. The only time
we convert them to milliseconds is when start_time and
end_time is sent as part of a query to TSDB.